+++ pkgs/jni/lib/src/lang/jdouble.dart
@@ -2,24 +2,18 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../jobject.dart';
-import '../types.dart';
-import 'jnumber.dart';
+import '../core_bindings.dart';
 
-final class _$JDouble$Type$ extends JType<JDouble> {
-  const _$JDouble$Type$();
-
-  @override
-  String get signature => r'Ljava/lang/Double;';
-}
-
-extension type JDouble._(JObject _$this) implements JNumber {
-  /// The type which includes information such as the signature of this class.
-  static const JType<JDouble> type = _$JDouble$Type$();
-
-  static final _class = JClass.forName(r'java/lang/Double');
-
-  static final _ctorId = _class.constructorId(r'(D)V');
-
-  JDouble(double num) : _$this = _ctorId<JDouble>(_class, [num]);
+extension JDoubleExtension on JDouble {
+  /// Returns the value as a Dart double.
+  ///
+  /// If [releaseOriginal] is true, the underlying reference is deleted
+  /// after conversion and this object will be marked as released.
+  double toDartDouble({bool releaseOriginal = false}) {
+    final ret = doubleValue();
+    if (releaseOriginal) {
+      release();
+    }
+    return ret;
+  }
 }