+++ pkgs/jni/lib/src/lang/jlong.dart
@@ -2,24 +2,18 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../jobject.dart';
-import '../types.dart';
-import 'jnumber.dart';
+import '../core_bindings.dart';
 
-final class _$JLong$Type$ extends JType<JLong> {
-  const _$JLong$Type$();
-
-  @override
-  String get signature => r'Ljava/lang/Long;';
-}
-
-extension type JLong._(JObject _$this) implements JNumber {
-  /// The type which includes information such as the signature of this class.
-  static const JType<JLong> type = _$JLong$Type$();
-
-  static final _class = JClass.forName(r'java/lang/Long');
-
-  static final _ctorId = _class.constructorId(r'(J)V');
-
-  JLong(int num) : _$this = _ctorId<JLong>(_class, [num]);
+extension JLongExtension on JLong {
+  /// Returns the value as a Dart int.
+  ///
+  /// If [releaseOriginal] is true, the underlying reference is deleted
+  /// after conversion and this object will be marked as released.
+  int toDartInt({bool releaseOriginal = false}) {
+    final ret = longValue();
+    if (releaseOriginal) {
+      release();
+    }
+    return ret;
+  }
 }