+++ pkgs/jni/tool/generate_jni_bindings.dart
@@ -7,15 +7,62 @@ import 'dart:io';
 import 'package:jnigen/jnigen.dart';
 import 'package:jnigen/src/elements/j_elements.dart' as j;
 
+// These core classes each have multiple constructors, such as a constructor
+// that converts a String to an Integer. We only want the constructor that takes
+// the primitive type, otherwise the APIs get messy. Filter out all other
+// constructors for these classes.
+const Map<String, String> _constructorAllowList = {
+  'Boolean': 'z',
+  'Byte': 'b',
+  'Character': 'c',
+  'Double': 'd',
+  'Float': 'f',
+  'Integer': 'i',
+  'Long': 'j',
+  'Short': 's',
+};
+
 class Renamer extends j.Visitor {
+  late j.ClassDecl _class;
+
   @override
   void visitClass(j.ClassDecl c) {
+    _class = c;
     c.name = 'J${c.originalName}';
+  }
+
+  @override
+  void visitMethod(j.Method m) {
+    if (!m.isConstructor) return;
+    final sig = _constructorAllowList[_class.originalName];
+    if (sig == null) return;
+    final lister = ListParams();
+    m.accept(lister);
+    m.isExcluded = !(lister.params.length == 1 && lister.params.first == sig);
+  }
+}
+
+class ListParams extends j.Visitor {
+  List<String> params = [];
+
+  @override
+  void visitParam(j.Param p) {
+    params.add(p.originalName);
   }
 }
 
 Future<void> main() async {
   final classes = [
+    'java.lang.Boolean',
+    'java.lang.Byte',
+    'java.lang.Character',
+    'java.lang.Double',
+    'java.lang.Float',
+    'java.lang.Integer',
+    'java.lang.Long',
+    'java.lang.Number',
+    'java.lang.Short',
+    'java.lang.String',
     'java.util.ArrayList',
     'java.util.Collection',
     'java.util.HashMap',