+++ pkgs/record_use/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.6.0-wip
+## 0.6.0
 
 - **Breaking Change**: Changed the JSON format and Dart API in various ways.